requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==6.0
django-environ==0.12.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.0.0
python-telegram-bot==22.5
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
